Added preliminary collision detection.
This version is broken in many ways, but it felt like time to commit.
Note that (at least h-collisions) appear to be about correct, but off by
some amount, which might be fixed with blit points as discussed on the
whiteboard.

darcs-hash:20040830043448-181b0-69e374f9bb2c4140526a8c2292e61896d646f0d1